requirements.txt
pandas
numpy
modelscope
datasets
torch
transformers
bitsandbytes
peft
trl
sentencepiece
protobuf
evaluate
nltk
ipykernel
accelerate
tqdm
nltk
tensorboard
unbabel-comet